+++ src/eca/server.clj
@@ -187,6 +187,9 @@
 (defmethod jsonrpc.server/receive-request "chat/open" [_ components params]
   (eventually (handlers/chat-open (with-config components) params)))
 
+(defmethod jsonrpc.server/receive-request "chat/history" [_ components params]
+  (eventually (handlers/chat-history (with-config components) params)))
+
 (defmethod jsonrpc.server/receive-notification "mcp/stopServer" [_ components params]
   (async-notify (handlers/mcp-stop-server (with-config components) params)))